Update source file structure.
