+++ Jigsaw-app/Jigsaw/Models/Game/GameCategory.swift
@@ -1,0 +1,18 @@
+//
+//  GameCategory.swift
+//  Jigsaw
+//
+//  Created by Ting Chen on 8/19/20.
+//  Copyright © 2020 DukeMobileDevCenter. All rights reserved.
+//
+
+import Foundation
+
+enum GameCategory: String, CaseIterable, Codable {
+    // Some dummy categories only for demo purposes
+    case immigration
+    case education
+    case housing
+    case medicare
+    case taxation
+}
