+++ Jigsaw-app/Jigsaw/Models/Question/Choice.swift
@@ -1,0 +1,14 @@
+//
+//  Choice.swift
+//  Jigsaw
+//
+//  Created by Ting Chen on 8/19/20.
+//  Copyright © 2020 DukeMobileDevCenter. All rights reserved.
+//
+
+import Foundation
+
+struct Choice: Codable {
+    let text: String
+    let value: String
+}
